KNMA-1741 Setup JWPlayer plugin for iOS
### Jira Ticket Link.

https://jiratl.kqed.org/browse/KNMA-1741

### Describe the feature or bug-fix.

This PR sets up the iOS/Swift plugin for the JWPlayer.
The CallbackMethod enum sets up messaging to Flutter so we can report back 
when certain actions are successful or we run into issues.
In the handle method we switch on the method type. Right now we are only 
doing an init and play method. Depending on the method, we will pull the values
out of the arguments that are passed in. 
We then call the functions with those args passed in which will either register the
license key or play the video.
The topViewController function gets the rootViewController which is where we 
present the VerticalPlayerViewController from.

In the example app's main.dart file, we simply have a text widget that shows the 
Flutter CallbackMethod strings and has a button that will make the player 
view controller appear. 
When the player appears it is set to auto play.

### Express how you feel about this PR with a GIF.

![](https://i.giphy.com/media/v1.Y2lkPTc5MGI3NjExdndzOHk0aDMycTkxMm8zc2VkOW9qdzR1OG14MWpyb3VmdmszNnd2aiZlcD12MV9pbnRlcm5hbF9naWZfYnlfaWQmY3Q9Zw/TRFcNpyLTdo2kUmkwm/giphy-downsized.gif)

### What areas of the App does it impact?

- Swift JWPlayer plugin setup
- Example app to test the plugin

### How to test? Be specific here. Call out any Environment or Accessibility pieces if applicable.

- Launch the app in Flutter or Xcode
- See that the CallbackMethod strings updates to confirm the license key was set
- Tap the `Play video` to present the player and watch the video

### Was documentation written or updated? If so, provide the link or links.

README will be updated in another PR to explain these changes.