KNMA-1720 Add title and description to video player - iOS only (#3)
### Jira Ticket Link.

https://jiratl.kqed.org/browse/KNMA-1720

### Describe the feature or bug-fix.

On iOS only:

This PR updates the the functions and method channel logic to take in
videoTitle and videoDescription properties. If the title or description
are nil or empty we set the isVisible property for the related one to
false.

We maybe update this later to allow for custom styling. The
JWPlayerSkinBuilder in their SDK allows for font and size changes.

### Express how you feel about this PR with a GIF.


![](https://i.giphy.com/media/v1.Y2lkPTc5MGI3NjExeHR5MTFkaHNzdHc5OWk0cDA4ZmtqNzN0N2o1OHZ1NjI5ZzU0bTJlZyZlcD12MV9pbnRlcm5hbF9naWZfYnlfaWQmY3Q9Zw/eg3f90cy2Mc6yVRWyR/giphy-downsized.gif)

### What areas of the App does it impact?

Displaying a String for title and String for description in the vertical
player

### How to test? Be specific here. Call out any Environment or
Accessibility pieces if applicable.

- Run the project in Flutter
- In example/lib/main.dart you can change the Strings at line 97/98
- Play a video and see the title appear
- Something seems to be broken with the example app and the description
is not showing up, but it works in our Flutter app. I will continue
trying to fix it later.

### Was documentation written or updated? If so, provide the link or
links.

N/A

![Simulator Screenshot - iPhone 15 Pro Max - 2024-08-27 at 08 54
32](https://github.com/user-attachments/assets/3660cf34-acaf-4d67-9a73-a9c924a72a9d)