KNMA-1822 Add ability to include captions on vertical videos
This PR adds new properties that can be passed through the method
channels that will allow captions to optionally be shown on the videos.

captionUrl is the URL where the captions file is located
captionLocale is the two letter language abbreviation
captionLanguageLabel is the full word of the language the captions are
in.

I included some commented code as hints to changing the position and
styling of the captions. We can remove that, but I'm pretty sure we
will want to customize it at some point. I figured it was easy to leave
it there than looking for it again in the docs.