KNMA-1822 Add ability to include captions on vertical videos (#5)
### Jira Ticket Link.

https://jiratl.kqed.org/browse/KNMA-1822

### Describe the feature or bug-fix.

Related API PR: https://github.com/KQED/mobile-app-server/pull/238
Related APP PR: TBD

This PR adds new properties that can be passed through the method
channels that will allow captions to optionally be shown on the videos.

captionUrl is the URL where the captions file is located captionLocale
is the two letter language abbreviation captionLanguageLabel is the full
word of the language the captions are in.

I included some commented code as hints to changing the position and
styling of the captions. We can remove that, but I'm pretty sure we will
want to customize it at some point. I figured it was easy to leave it
there than looking for it again in the docs.

### Express how you feel about this PR with a GIF.


![](https://i.giphy.com/media/v1.Y2lkPTc5MGI3NjExN2UwOGhpbDlzMGtiNmJzcTdrN2dqMjhiNWFmMXI1dDRtNDlrNXV2MyZlcD12MV9pbnRlcm5hbF9naWZfYnlfaWQmY3Q9Zw/6znpHNXfBQJnSWhwmm/giphy-downsized.gif)


### What areas of the Plugin does it impact?

- Adding ability to show captions on Vertical Videos

### How to test? Be specific here. Call out any Environment or
Accessibility pieces if applicable.

- Open project in Vscode
- Run the example app
- Hit the Play Video button
- Tap the captions icon and select English to show the captions

### Was documentation written or updated? If so, provide the link or
links.

https://kb.kqed.org/display/ProDoc/Vertical+Video+-+JW+Player